backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
python-dotenv>=1.0.1
pydantic>=2.9.0,<3.0.0
pydantic-settings>=2.5.0,<3.0.0
openai>=1.50.0,<2.0.0
chromadb>=0.5.0,<1.0.0
google-auth>=2.34.0,<3.0.0
google-auth-oauthlib>=1.2.0,<2.0.0
google-api-python-client>=2.146.0,<3.0.0
python-multipart>=0.0.12
python-jose[cryptography]>=3.3.0,<4.0.0
httpx>=0.27.0,<1.0.0
PyPDF2>=3.0.1,<4.0.0
